Supports dense tensor accesses of arbitrary order

+++ src/domain.cpp
@@ -58,8 +58,8 @@ std::ostream &operator<<(std::ostream &os, const IndexSet &is) {
 
 
 // class IndexDomain
-size_t IndexDomain::getSize() const {
-  size_t size = 1;
+int IndexDomain::getSize() const {
+  int size = 1;
   for (auto &indexSet : getIndexSets()) {
     size *= indexSet.getSize();
   }
